会社概要をフッターから除く
